Move shared load and build methods to module
